Create description.txt
Problem Statement:
The sinking of the Titanic is one of the most infamous shipwrecks in history.
On April 15, 1912, during her maiden voyage, the widely considered “unsinkable” RMS Titanic sank
after colliding with an iceberg. Unfortunately, there weren’t enough lifeboats for everyone on board,
resulting in the death of 1502 out of 2224 passengers and crew.
Use the Titanic dataset to build a model that predicts whether a passenger on the Titanic survived or
not. This is a classic beginner project with readily available data.
The dataset typically used for this project contains information about individual passengers, such as
their age, gender, ticket class, fare, cabin, and whether or not they survived.